Sketch the image/requirement.txt
streamlit==0.89.0
opencv-python==4.5.3.56
pillow==8.4.0
matplotlib==3.4.3